# openapi.json
{"swagger":"2.0","info":{"version":"9.6.1","title":"@fastify/swagger"},"definitions":{},"paths":{"/healthz":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/readyz":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/orders/":{"post":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["order_number","order_date","delivery_date","buyer","recipient","delivery_method_id","payment_method_id","order_status_id","shipping_cost","items"],"properties":{"order_number":{"minLength":1,"maxLength":50,"type":"string"},"order_date":{"format":"date-time","type":"string"},"delivery_date":{"format":"date-time","type":"string"},"buyer":{"type":"object","required":["id","phone","address"],"properties":{"id":{"type":"integer"},"phone":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}},"address":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}}}},"recipient":{"type":"object","required":["id","phone","address"],"properties":{"id":{"type":"integer"},"phone":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}},"address":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}}}},"delivery_method_id":{"type":"integer"},"payment_method_id":{"type":"integer"},"order_status_id":{"type":"integer"},"shipping_cost":{"minimum":0,"maximum":1000000000,"type":"integer"},"note":{"maxLength":500,"type":"string"},"items":{"minItems":1,"maxItems":100,"type":"array","items":{"type":"object","required":["item_id","quantity"],"properties":{"item_id":{"type":"integer"},"quantity":{"minimum":1,"maximum":10000,"type":"integer"}}}}}}}],"responses":{"201":{"description":"Default Response","schema":{"type":"object","required":["id","order_number","order_date","delivery_date","buyer_id","buyer_name","buyer_phone","buyer_address","recipient_id","recipient_name","recipient_phone","recipient_address","delivery_method_id","payment_method_id","order_status_id","shipping_cost","subtotal_amount","total_amount","note","created_at","updated_at"],"properties":{"id":{"type":"integer"},"order_number":{"minLength":1,"maxLength":50,"type":"string"},"order_date":{"format":"date-time","type":"string"},"delivery_date":{"format":"date-time","type":"string"},"buyer_id":{"type":"integer"},"buyer_name":{"minLength":1,"maxLength":255,"type":"string"},"buyer_phone":{"anyOf":[{"maxLength":25,"type":"string"},{"type":"null"}]},"buyer_address":{"anyOf":[{"maxLength":500,"type":"string"},{"type":"null"}]},"recipient_id":{"type":"integer"},"recipient_name":{"minLength":1,"maxLength":255,"type":"string"},"recipient_phone":{"anyOf":[{"maxLength":25,"type":"string"},{"type":"null"}]},"recipient_address":{"anyOf":[{"maxLength":500,"type":"string"},{"type":"null"}]},"delivery_method_id":{"type":"integer"},"payment_method_id":{"type":"integer"},"order_status_id":{"type":"integer"},"shipping_cost":{"minimum":0,"maximum":1000000000,"type":"integer"},"subtotal_amount":{"minimum":0,"maximum":1000000000,"type":"integer"},"total_amount":{"minimum":0,"maximum":1000000000,"type":"integer"},"note":{"anyOf":[{"type":"string"},{"type":"null"}]},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}}}}}},"get":{"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","required":["id","order_number","order_date","delivery_date","recipient_name","recipient_address","order_status_name","total_amount"],"properties":{"id":{"type":"integer"},"order_number":{"minLength":1,"maxLength":50,"type":"string"},"order_date":{"format":"date-time","type":"string"},"delivery_date":{"format":"date-time","type":"string"},"recipient_name":{"minLength":1,"maxLength":255,"type":"string"},"recipient_address":{"anyOf":[{"maxLength":500,"type":"string"},{"type":"null"}]},"order_status_name":{"type":"string"},"total_amount":{"minimum":0,"maximum":1000000000,"type":"integer"}}}}}}}},"/api/orders/{id}":{"put":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["order_number","order_date","delivery_date","buyer","recipient","delivery_method_id","payment_method_id","order_status_id","shipping_cost","items"],"properties":{"order_number":{"minLength":1,"maxLength":50,"type":"string"},"order_date":{"format":"date-time","type":"string"},"delivery_date":{"format":"date-time","type":"string"},"buyer":{"type":"object","required":["id","phone","address"],"properties":{"id":{"type":"integer"},"phone":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}},"address":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}}}},"recipient":{"type":"object","required":["id","phone","address"],"properties":{"id":{"type":"integer"},"phone":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}},"address":{"type":"object","required":["id"],"properties":{"id":{"type":"integer"}}}}},"delivery_method_id":{"type":"integer"},"payment_method_id":{"type":"integer"},"order_status_id":{"type":"integer"},"shipping_cost":{"minimum":0,"maximum":1000000000,"type":"integer"},"note":{"maxLength":500,"type":"string"},"items":{"minItems":1,"maxItems":100,"type":"array","items":{"type":"object","required":["item_id","quantity"],"properties":{"item_id":{"type":"integer"},"quantity":{"minimum":1,"maximum":10000,"type":"integer"}}}}}}},{"type":"number","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response","schema":{"type":"object","required":["id","order_number","order_date","delivery_date","buyer_id","buyer_name","buyer_phone","buyer_address","recipient_id","recipient_name","recipient_phone","recipient_address","delivery_method_id","payment_method_id","order_status_id","shipping_cost","subtotal_amount","total_amount","note","created_at","updated_at"],"properties":{"id":{"type":"integer"},"order_number":{"minLength":1,"maxLength":50,"type":"string"},"order_date":{"format":"date-time","type":"string"},"delivery_date":{"format":"date-time","type":"string"},"buyer_id":{"type":"integer"},"buyer_name":{"minLength":1,"maxLength":255,"type":"string"},"buyer_phone":{"anyOf":[{"maxLength":25,"type":"string"},{"type":"null"}]},"buyer_address":{"anyOf":[{"maxLength":500,"type":"string"},{"type":"null"}]},"recipient_id":{"type":"integer"},"recipient_name":{"minLength":1,"maxLength":255,"type":"string"},"recipient_phone":{"anyOf":[{"maxLength":25,"type":"string"},{"type":"null"}]},"recipient_address":{"anyOf":[{"maxLength":500,"type":"string"},{"type":"null"}]},"delivery_method_id":{"type":"integer"},"payment_method_id":{"type":"integer"},"order_status_id":{"type":"integer"},"shipping_cost":{"minimum":0,"maximum":1000000000,"type":"integer"},"subtotal_amount":{"minimum":0,"maximum":1000000000,"type":"integer"},"total_amount":{"minimum":0,"maximum":1000000000,"type":"integer"},"note":{"anyOf":[{"type":"string"},{"type":"null"}]},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}}}}}}},"/api/items/":{"post":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["name","price"],"properties":{"name":{"minLength":1,"maxLength":100,"type":"string"},"price":{"minimum":0,"maximum":1000000000,"type":"integer"}}}}],"responses":{"201":{"description":"OK","schema":{"description":"OK"}}}},"get":{"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","required":["id","name","price"],"properties":{"id":{"type":"integer"},"name":{"minLength":1,"maxLength":100,"type":"string"},"price":{"minimum":0,"maximum":1000000000,"type":"integer"}}}}}}}},"/api/items/{id}":{"get":{"parameters":[{"type":"number","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response","schema":{"type":"object","required":["id","name","price"],"properties":{"id":{"type":"integer"},"name":{"minLength":1,"maxLength":100,"type":"string"},"price":{"minimum":0,"maximum":1000000000,"type":"integer"}}}}}},"put":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["name","price"],"properties":{"name":{"minLength":1,"maxLength":100,"type":"string"},"price":{"minimum":0,"maximum":1000000000,"type":"integer"}}}},{"type":"number","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"OK","schema":{"description":"OK"}}}}},"/api/order-statuses/":{"get":{"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer"},"name":{"type":"string"}}}}}}}},"/api/payment-methods/":{"get":{"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer"},"name":{"type":"string"}}}}}}}},"/api/delivery-methods/":{"get":{"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer"},"name":{"type":"string"}}}}}}}},"/api/persons/":{"post":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["name"],"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"}}}}],"responses":{"201":{"description":"OK","schema":{"description":"OK"}}}}},"/api/persons/search":{"get":{"parameters":[{"minLength":1,"type":"string","required":true,"in":"query","name":"name"}],"responses":{"200":{"description":"Default Response","schema":{"maxItems":50,"type":"array","items":{"type":"object","required":["id","name"],"properties":{"id":{"type":"integer"},"name":{"type":"string"}}}}}}}},"/api/phones/":{"post":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["person_id","phone_number"],"properties":{"person_id":{"type":"integer"},"phone_number":{"minLength":1,"maxLength":25,"type":"string"}}}}],"responses":{"201":{"description":"OK","schema":{"description":"OK"}}}}},"/api/phones/search":{"get":{"parameters":[{"type":"integer","required":true,"in":"query","name":"person_id"},{"minLength":1,"type":"string","required":true,"in":"query","name":"phone_number"}],"responses":{"200":{"description":"Default Response","schema":{"maxItems":50,"type":"array","items":{"type":"object","required":["id","phone_number"],"properties":{"id":{"type":"integer"},"phone_number":{"minLength":1,"maxLength":25,"type":"string"}}}}}}}},"/api/addresses/":{"post":{"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["person_id","address"],"properties":{"person_id":{"type":"integer"},"address":{"minLength":1,"maxLength":500,"type":"string"}}}}],"responses":{"201":{"description":"OK","schema":{"description":"OK"}}}}},"/api/addresses/search":{"get":{"parameters":[{"type":"integer","required":true,"in":"query","name":"person_id"},{"minLength":1,"type":"string","required":true,"in":"query","name":"address"}],"responses":{"200":{"description":"Default Response","schema":{"maxItems":50,"type":"array","items":{"type":"object","required":["id","address"],"properties":{"id":{"type":"integer"},"address":{"minLength":1,"maxLength":500,"type":"string"}}}}}}}}}}